report

+++ payrallproject/Models/Domains/OT.cs
@@ -7,7 +7,7 @@ namespace payrallproject.Models.Domains
         [Key]
         public int? Id { get; set; }
         public string? Name { get; set; }
-        public int? Rate { get; set; }
+        public decimal? Rate { get; set; }
         public bool? IsActive { get; set; } = true;
     }
 }
