+++ payrallproject/Models/Dtos/OTDto.cs
@@ -3,7 +3,7 @@
     public class OTDto
     {
         public string? Name { get; set; }
-        public int? Rate { get; set; }
+        public decimal? Rate { get; set; }
         public bool? IsActive { get; set; }
     }
 }
